More core logic changes. Major switch to MySQL

+++ keno/sql/tables/drawings.sql
@@ -1,8 +1,8 @@
 CREATE TABLE `drawings` (
     `id` INT UNSIGNED NOT NULL AUTO_INCREMENT,
     `date` DATETIME,
-    `high_bits` INT UNSIGNED NOT NULL,
-    `low_bits` INT UNSIGNED NOT NULL,
+    `high_bits` BIGINT UNSIGNED NOT NULL,
+    `low_bits` BIGINT UNSIGNED NOT NULL,
     `number_string` TEXT,
     PRIMARY KEY(`id`)
 );
